lab-2-web/backend/requirements.txt
asgiref==3.7.2
Django
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
